Refactor redemption modules
