+++ src/redemptions/commands/update_redeemable_positions.py
@@ -20,11 +20,7 @@ from src.common.clients import (
     get_execution_client,
     setup_clients,
 )
-from src.common.contracts import (
-    Erc20Contract,
-    MulticallContract,
-    os_token_redeemer_contract,
-)
+from src.common.contracts import Erc20Contract, MulticallContract
 from src.common.logging import LOG_LEVELS, setup_logging
 from src.common.startup_check import (
     check_execution_nodes_network,
@@ -42,6 +38,7 @@ from src.redemptions.api_client import (
     RABBY_API_SOURCE,
     APIClient,
 )
+from src.redemptions.contracts import os_token_redeemer_contract
 from src.redemptions.graph import (
     graph_get_leverage_positions,
     graph_get_os_token_holders,
