+++ src/redemptions/tasks.py
@@ -7,10 +7,10 @@ from sw_utils import OsTokenConverter
 from sw_utils.typings import ChainHead, ProtocolConfig
 from web3.types import Wei
 
-from src.common.contracts import os_token_redeemer_contract
 from src.common.execution import get_finalized_block_number
 from src.common.protocol_config import get_protocol_config
 from src.config.settings import settings
+from src.redemptions.contracts import os_token_redeemer_contract
 from src.redemptions.fetch_positions import (
     fetch_positions_with_processed_shares,
     update_positions_cache,
